+++ tests/fake-tasks/src/fake_tasks/task_list_no_authors.py
@@ -1,12 +1,6 @@
 from fractal_task_tools.task_models import NonParallelTask
 
 DOCS_LINK = "https://www.example.org"
-INPUT_MODELS = [
-    ("fake_tasks", "models.py", "ModelMixedDocstrings"),
-    ("fake_tasks", "task2_with_inheritance.py", "GenericParameters"),
-    ("fake_tasks", "task2_with_inheritance.py", "SpecificParameters"),
-    ("fake_tasks", "task3_for_descriptions.py", "ParametersA"),
-]
 TASK_LIST = [
     NonParallelTask(
         name="Task1",