+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # Unreleased
 
 * Support getting manifest `authors` from `pyproject.toml` in the current directory (\#136).
+* Drop support for `INPUT_MODELS`, in favor of new style for class-attribute docstrings (\#137).
 
 # 0.4.0